+++ src/indicators/IndicatorVariables.cpp
@@ -4,10 +4,11 @@
 using namespace geode::prelude;
 using namespace IndicatorVars;
 
-void IndicatorVars::updateSettings() {
+void IndicatorVars::updateSettings() { // kinda randomly sorted but its an organised mess so good enough
     auto mod = Mod::get();
     onlySelected = mod->getSettingValue<bool>("only-selected");
     onlyTriggers = mod->getSettingValue<bool>("only-triggers");
+    onlySpawn = mod->getSettingValue<bool>("only-spawn");
     
     auto maxDistanceSetting = mod->getSettingValue<double>("max-distance");
     maxDistance = maxDistanceSetting * maxDistanceSetting;
@@ -28,6 +29,8 @@ void IndicatorVars::updateSettings() {
     groupBlacklist = parseIntArray(mod->getSettingValue<std::string>("group-blacklist"));
     triggerBlacklist = parseIntArray(mod->getSettingValue<std::string>("trigger-blacklist") 
     + "," + mod->getSettingValue<std::string>("forced-trigger-blacklist")); // saves checking 2 diff sets
+    
+    // spawnIndicators = mod->getSettingValue<bool>("spawn-indicators");
     
     settingsButtonTexture = "nwo5.trigger_indicators_v2/" + mod->getSettingValue<std::string>("settings-button-texture") + ".png";